Module Topkg.Exts

Exts defines sets of file extensions.

File extensions

type ext

The type for file extensions.

type t = ext list

The type for lists of file extensions.

val interface : t

interface is exts [".mli"; ".cmi"; ".cmti"].

val cmx : ext list

cmx is ext ".cmx".

val api : t

api is interface @ cmx.

val real_c_library : ext list

real_c_library is the extension for C libraries (archives). This should be used by C libraries (e.g. stubs) compiled by OCaml. For example Topkg.Pkg.clib uses this. The actual value is determined from OCaml's configuration.

val c_library : ext list

c_library is the extension for C libraries (archives). This is like real_c_library but for those C archive that are generated by OCaml build artefacts. The actual value is determined from OCaml's configuration.

val c_dll_library : ext list

c_dll_library is the extension for C dynamic libraries (archives). The actual value is determined from OCaml's configuration.

val library : ext list

library is exts [".cma"; ".cmxa"; ".cmxs"] @ c_library

val module_library : ext list

module_library is (api @ library).

val exe : ext list

exe is the extension for executables. The actual value is determined from OCaml's configuration.

val exts : string list -> ext list

exts ss is ss as a list of extensions.

val ext : string -> ext list

ext s is s as a list of extensions.