Module OS.Env

Environment variables

Variables

val var : string -> string option

var name is the value of the environment variable name, if defined.

val opt_var : string -> absent:string -> string

opt_var name ~absent is the value of the optionally defined environment variable name, if defined, and absent if undefined.