Module Topkg.Private

Private definitions.

Warning. The following definitions are subject to change even between minor versions of the library. Topkg users must not use these definitions to describe their package.

Private

val disable_main : unit -> unit

disable_main () disables Topkg's main invoked on Pkg.describe. Invoke this function in your main function if you are not using Topkg in a description file but as as a library.

module Codec : sig ... end

Topkg interprocess communication codec.

module Pkg : sig ... end

Package description.

module Ipc : sig ... end

Topkg interprocess communication.

module Opam : sig ... end

opam helpers.